Extended GP Distributions

Distributions

The extended generalized Pareto families proposed in Naveau et al. (2016) and the extended GP model developed by Gamet and Jalbert (2020) are implemented and ready to be used just like any UnivariateDistribution from the Distributions.jl package.

ExtendedExtremes.EGPpower — Type

EGPpower(σ, ξ, κ)

EGPpower corresponds to the first extended GP model of Naveau et al. (2016), with the power law distribution G(v) = v^κ.

It is a three parameters family: κ controls the shape of the lower tail, σ is a scale parameter, and ξ controls the rate of upper tail decay.

EGPpower(σ, ξ, κ)   # EGP of Naveau et al. (2016) (type 1) with scale parameter σ, rate of upper tail decay ξ and shape of the lower tail κ.

params(d)           # Get the parameters, i.e. (σ, ξ, κ)

Reference :

Naveau, P., Huser, R., Ribereau, P., and Hannart, A. (2016), Modeling jointly low, moderate, and heavy rainfall intensities without a threshold selection, Water Resour. Res., 52, 2753– 2769, doi:10.1002/2015WR018552.

ExtendedExtremes.EGPpowermix — Type

EGPpowermix(σ, ξ, κ₁, κ₂, p)

EGPpowermix corresponds to the second extended GP model of Naveau et al. (2016), which is a mixture of power laws with G(v) = pv^κ₁ + (1-p)v^κ₂.

It is a five parameters family: κ₁ controls the lower tail behavior, κ₂ modifies the shape of the density in its central part, σ is a scale parameter, and ξ is a shape parameter.

EGPpowermix(σ, ξ, κ₁, κ₂, p)   # EGP of Naveau et al. (2016) (type 2) with parameters σ, ξ, κ₁, κ₂, p.

params(d)           # Get the parameters, i.e. (σ, ξ, κ₁, κ₂, p)

Reference :

Naveau, P., Huser, R., Ribereau, P., and Hannart, A. (2016), Modeling jointly low, moderate, and heavy rainfall intensities without a threshold selection, Water Resour. Res., 52, 2753– 2769, doi:10.1002/2015WR018552.

ExtendedExtremes.EGPbeta — Type

EGPbeta(σ, ξ, δ)

EGPbeta corresponds to the third extended GP model of Naveau et al. (2016).

It is a three parameters family: δ is a threshold tuning parameter, σ is a scale parameter, and ξ is a shape parameter.

EGPbeta(σ, ξ, δ)   # EGP of Naveau et al. (2016) (type 3) with parameters σ, ξ, δ.

params(d)           # Get the parameters, i.e. (σ, ξ, δ)

Reference :

Naveau, P., Huser, R., Ribereau, P., and Hannart, A. (2016), Modeling jointly low, moderate, and heavy rainfall intensities without a threshold selection, Water Resour. Res., 52, 2753– 2769, doi:10.1002/2015WR018552.

ExtendedExtremes.EGPbetapower — Type

EGPbetapower(σ, ξ, δ)

EGPbetapower corresponds to the fourth extended GP model of Naveau et al. (2016).

It is a four parameters family: δ is a threshold tuning parameter, κ is a parameter that controls the lower tail behavior, σ is a scale parameter, and ξ is a shape parameter.

EGPbetapower(σ, ξ, δ, κ)   # EGP of Naveau et al. (2016) (type 4) with parameters σ, ξ, δ, κ.

params(d)           # Get the parameters, i.e. (σ, ξ, δ, κ)

Reference :

Naveau, P., Huser, R., Ribereau, P., and Hannart, A. (2016), Modeling jointly low, moderate, and heavy rainfall intensities without a threshold selection, Water Resour. Res., 52, 2753– 2769, doi:10.1002/2015WR018552.

ExtendedExtremes.EGPnormal — Type

EGPnormal(σ, ξ, κ)

EGPnormal corresponds to the extended GP model developed by Gamet and Jalbert (2020), with the truncated normal distribution of mean 1 and variance 1/κ^2.

It is a three parameters family: κ controls the shape of the lower tail, σ is a scale parameter, and ξ controls the rate of upper tail decay.

EGPnormal(σ, ξ, κ)   # EGP of Gamet and Jalbert (2020) with scale parameter σ, rate of upper tail decay ξ and shape of the lower tail κ.

params(d)           # Get the parameters, i.e. (σ, ξ, κ)

Reference :

Unpublished for now

Common Interface

A series of methods are implemented for each EGPD, which provide useful functionalities such as moment computation, pdf evaluation, and sampling (i.e. random number generation).

Probability Evaluation

Distributions.cdf — Method

cdf(d::UnivariateDistribution, x::Real)

Evaluate the cumulative probability at x.

See also ccdf, logcdf, and logccdf.

Distributions.logcdf — Method

logcdf(d::UnivariateDistribution, x::Real)

The logarithm of the cumulative function value(s) evaluated at x, i.e. log(cdf(x)).

Distributions.logpdf — Method

logpdf(d::UnivariateDistribution, x::Real)

Evaluate the logarithm of probability density (mass) at x. Whereas there is a fallback implemented logpdf(d, x) = log(pdf(d, x)). Relying on this fallback is not recommended in general, as it is prone to overflow or underflow.

Distributions.pdf — Method

pdf(d::UnivariateDistribution, x::Real)

Evaluate the probability density (mass) at x.

See also: logpdf.

Statistics.quantile — Method

quantile(d::UnivariateDistribution, q::Real)

Evaluate the inverse cumulative distribution function at q.

See also: cquantile, invlogcdf, and invlogccdf.

Computation of statistics

Distributions.insupport — Method

insupport(d::UnivariateDistribution, x::Any)

When x is a scalar, it returns whether x is within the support of d (e.g., insupport(d, x) = minimum(d) <= x <= maximum(d)). When x is an array, it returns whether every element in x is within the support of d.

Generic fallback methods are provided, but it is often the case that insupport can be done more efficiently, and a specialized insupport is thus desirable. You should also override this function if the support is composed of multiple disjoint intervals.

Base.maximum — Method

maximum(d::UnivariateDistribution)

Return the maximum of the support of d.

Base.minimum — Method

minimum(d::UnivariateDistribution)

Return the minimum of the support of d.

Parameter Retrieval

StatsBase.params — Method

params(d::UnivariateDistribution)

Return a tuple of parameters. Let d be a distribution of type D, then D(params(d)...) will construct exactly the same distribution as $d$.

Index

ExtendedExtremes.EGPbeta

ExtendedExtremes.EGPbetapower

ExtendedExtremes.EGPnormal

ExtendedExtremes.EGPpower

ExtendedExtremes.EGPpowermix

Base.maximum

Base.minimum

Distributions.cdf

Distributions.insupport

Distributions.logcdf

Distributions.logpdf

Distributions.pdf

Statistics.quantile

StatsBase.params