Functions

ExtendedExtremes.BIC — Method

BIC(fm::ContinuousUnivariateDistribution, data::Array{<:Real,1}; censoring::Real=0)

Calculates the Bayesian information criterion (BIC) of the specified model.

\[\textrm{BIC} = \ln f_{(Y|\hat{\theta}_{MV})}(y) - \frac{k}{2}\ln{n}\]

ExtendedExtremes.BIC — Method

BIC(fm, data, covariate)

ExtendedExtremes.EGPbetafit — Method

EGPbetafit()

Estimate the third extended GP model of Naveau et al. (2016) parameters by maximum likelihood.

ExtendedExtremes.EGPbetapowerfit — Method

EGPbetapowerfit()

Estimate the fourth extended GP model of Naveau et al. (2016) parameters by maximum likelihood.

ExtendedExtremes.EGPnormalfit — Method

EGPnormalfit()

Estimate the extended GP model of Gamet and Jalbert (2020) parameters by maximum likelihood.

ExtendedExtremes.EGPpowerfit — Method

EGPpowerfit()

Estimate the first extended GP model of Naveau et al. (2016) parameters by maximum likelihood.

ExtendedExtremes.EGPpowermixfit — Method

EGPpowermixfit()

Estimate the second extended GP model of Naveau et al. (2016) parameters by maximum likelihood.

ExtendedExtremes.T_transform — Method

T_transform(m::CDFT, x::Real)

ExtendedExtremes.T_transform — Method

T_transform(Yc, Xc, Xp, x) (stationnary)

ExtendedExtremes.T_transform — Method

T_transform(Yx, Xc, Xc_cov, Xp, Xp_cov, x)  (non-stationnary)

ExtendedExtremes.bias_correct — Method

bias_correct(obs, ref, covariate_ref; fut, covariate_fut, censoring_obs, censoring_sim, wet_thresh)

ExtendedExtremes.index_R10 — Method

index_R10(ts::TimeArray)

Returns the relative frequency of days with precipitation above 10 mm.

ExtendedExtremes.index_RV20max — Method

index_RV20max(ts::TimeArray)

Returns the 20-years return value of heavy precipitation.

ExtendedExtremes.index_RVxmax — Method

index_RVxmax(ts::TimeArray, p::Real)

Returns the p-years return value of heavy precipitation.

ExtendedExtremes.index_Rx — Method

index_Rx(ts::TimeArray, x::Real)

Returns the relative frequency of days with precipitation above x mm.

ExtendedExtremes.index_SDII — Method

index_SDII(ts::TimeArray)

Returns the precipitation amount in wet days divided by the number of wet days.

ExtendedExtremes.index_p90wet — Method

index_p90wet(ts::TimeArray)

Returns the 90% percentile of wet days considering days with precipitation above 1 mm.

ExtendedExtremes.index_p98wet — Method

index_p98wet(ts::TimeArray)

Returns the 98% percentile of wet days considering days with precipitation above 1 mm.

ExtendedExtremes.index_p98wetamount — Method

index_p98wetamount(ts::TimeArray)

Returns the precipitation amount from days above 98% percentile.

ExtendedExtremes.index_pXwet — Method

index_pXwet(ts::TimeArray, p::Real; wet_thresh::Real=1.0)

Returns the p percentile of wet days considering days with precipitation above wet_thresh (mm).

ExtendedExtremes.index_pXwetamount — Method

index_pXwetamount(ts::TimeArray, p::Real; wet_thresh=1.0)

Returns the precipitation amount from days above p percentile.

ExtendedExtremes.index_skew — Method

index_skew(ts::TimeArray)

Returns the skewness of ts.

ExtendedExtremes.ppbuild — Method

ppbuild()

ExtendedExtremes.ppbuild — Method

ppbuild() (standardized)

ExtendedExtremes.qqbuild — Method

qqbuild()

ExtendedExtremes.qqbuild — Method

qqbuild() (standardized)

Extremes.diagnosticplots — Method

diagnosticplots(fm::UnivariateDistribution, data::AbstractVector)

Extremes.histplot — Method

histplot(fm::UnivariateDistribution, data::AbstractVector)

Histogram plot

Extremes.probplot — Method

probplot()

Probability plot

Extremes.qqplot — Method

qqplot()

Quantile-Quantile Plot

Extremes.returnlevelplot — Method

returnlevelplot(fm::UnivariateDistribution, data::AbstractVector)

Return level plot

ExtendedExtremes.CDFT — Type

CDFT(Yc, Xc, Xp)

ExtendedExtremes.EGPpower — Method

EGPpower(fm::nonstatEGPpower{T}, covariate::Real)

Pour obtenir le modèle stationnaire à partir de la valeur de la variable explicative.